- moved server directory under /src/ - installed typeorm and initialized postgres template - installed and integrated typescript
