+++ .gitignore
@@ -1,2 +1,6 @@
-node_modules
-.DS_Store
+.idea/
+.vscode/
+node_modules/
+build/
+tmp/
+temp/
