+++ package.json
@@ -1,36 +1,43 @@
 {
-  "name": "bonsai-interview",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "dependencies": {
-    "apollo-boost": "^0.4.3",
-    "apollo-server-express": "^2.7.0",
-    "bootstrap": "^4.3.1",
-    "express": "^4.17.1",
-    "graphql": "^14.4.2",
-    "react": "^16.8.6",
-    "react-apollo": "^2.5.8",
-    "react-dom": "^16.8.6",
-    "reactstrap": "^8.0.1"
-  },
-  "scripts": {
-    "server": "nodemon ./server/server.js",
-    "client": "webpack-dev-server --mode development --open",
-    "dev": "concurrently \"yarn client\" \"yarn server\""
-  },
-  "devDependencies": {
-    "@babel/core": "^7.5.5",
-    "@babel/plugin-proposal-decorators": "^7.4.4",
-    "@babel/preset-react": "^7.0.0",
-    "babel-loader": "^8.0.6",
-    "concurrently": "^4.1.1",
-    "css-loader": "^3.1.0",
-    "html-webpack-plugin": "^3.2.0",
-    "nodemon": "^1.19.1",
-    "style-loader": "^0.23.1",
-    "webpack": "^4.37.0",
-    "webpack-cli": "^3.3.6",
-    "webpack-dev-server": "^3.7.2"
-  }
+   "name": "bonsai-interview",
+   "version": "1.0.0",
+   "main": "index.js",
+   "license": "MIT",
+   "dependencies": {
+      "apollo-boost": "^0.4.3",
+      "apollo-server-express": "^2.7.0",
+      "bootstrap": "^4.3.1",
+      "express": "^4.17.1",
+      "graphql": "^14.4.2",
+      "react": "^16.8.6",
+      "react-apollo": "^2.5.8",
+      "react-dom": "^16.8.6",
+      "reactstrap": "^8.0.1",
+      "typeorm": "0.2.32",
+      "reflect-metadata": "^0.1.10",
+      "pg": "^8.4.0"
+   },
+   "scripts": {
+      "server": "nodemon -w src --ext ts --exec ts-node ./src/server/server.ts",
+      "client": "webpack-dev-server --mode development --open",
+      "dev": "concurrently \"yarn client\" \"yarn server\"",
+      "start": "ts-node ./src/server/server.ts"
+   },
+   "devDependencies": {
+      "@babel/core": "^7.5.5",
+      "@babel/plugin-proposal-decorators": "^7.4.4",
+      "@babel/preset-react": "^7.0.0",
+      "@types/node": "^8.10.66",
+      "babel-loader": "^8.0.6",
+      "concurrently": "^4.1.1",
+      "css-loader": "^3.1.0",
+      "html-webpack-plugin": "^3.2.0",
+      "nodemon": "^1.19.1",
+      "style-loader": "^0.23.1",
+      "ts-node": "3.3.0",
+      "typescript": "^3.3.3333",
+      "webpack": "^4.37.0",
+      "webpack-cli": "^3.3.6",
+      "webpack-dev-server": "^3.7.2"
+   }
 }
