+++ src/server/modules/merchant/resolvers.ts
@@ -1,4 +1,5 @@
-const { merchants } = require("../../../mockMerchantData")
+const { merchants } = require("../../mockMerchantData")
+
 
 const resolvers = {
   Query: {
